GH-920: Fix Class Tangle
- move `EndpointHandlerMethod` to its own file
- remove backward reference from `...Configuration` to its builder
- fix docs
