+++ spring-kafka/src/test/java/org/springframework/kafka/retrytopic/RetryTopicConfigurerTests.java
@@ -110,7 +110,7 @@ class RetryTopicConfigurerTests {
 	private RetryTopicConfiguration.TopicCreation topicCreationConfig;
 
 	@Mock
-	private RetryTopicConfigurer.EndpointHandlerMethod endpointHandlerMethod;
+	private EndpointHandlerMethod endpointHandlerMethod;
 
 	@Mock
 	private ConsumerRecord<?, ?> consumerRecordMessage;
@@ -317,7 +317,7 @@ class RetryTopicConfigurerTests {
 		// setup
 		NoOpsClass noOps = new NoOpsClass();
 		willReturn(noOps).given(beanFactory).getBean(NoOpsClass.class);
-		RetryTopicConfigurer.EndpointHandlerMethod handlerMethod =
+		EndpointHandlerMethod handlerMethod =
 				RetryTopicConfigurer.createHandlerMethodWith(NoOpsClass.class, noOpsMethodName);
 
 		// given
@@ -335,7 +335,7 @@ class RetryTopicConfigurerTests {
 		String beanName = NoOpsClass.class.getSimpleName() + "-handlerMethod";
 		given(defaultListableBeanFactory.getBean(beanName)).willReturn(new NoOpsClass());
 		willThrow(NoSuchBeanDefinitionException.class).given(defaultListableBeanFactory).getBean(NoOpsClass.class);
-		RetryTopicConfigurer.EndpointHandlerMethod handlerMethod =
+		EndpointHandlerMethod handlerMethod =
 				RetryTopicConfigurer.createHandlerMethodWith(NoOpsClass.class, noOpsMethodName);
 
 		// given
